% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HpolytopeSparseClass.R
\docType{class}
\name{HpolytopeSparse-class}
\alias{HpolytopeSparse-class}
\alias{HpolytopeSparse}
\title{An R class to represent an H-polytope defined by a sparse matrix}
\description{
A sparse H-polytope is a convex polytope defined by a set of linear inequalities or equivalently a
\eqn{d}-dimensional H-polytope with \eqn{m} facets is defined by a \eqn{m\times d} matrix A and
a \eqn{m}-dimensional vector b, s.t.: \eqn{Ax\leq b}, where \eqn{A} is sparse.
}
\details{
\describe{
   \item{Aineq}{An \eqn{m\times d} sparse numerical matrix for the inequalities.}

   \item{bineq}{An \eqn{m}-dimensional vector for the ineqaulities.}

   \item{Aineq}{An \eqn{m'\times d} sparse numerical matrix for the equalities.}

   \item{bineq}{An \eqn{m'}-dimensional vector for the eqaulities.}

   \item{lb}{\eqn{d}-dimensional vector lb.}

   \item{ub}{\eqn{d}-dimensional vector ub.}

   \item{type}{A character with default value 'HpolytopeSparse', to declare the representation of the polytope.}
}
}
\value{
An object of class \code{HpolytopeSparse} representing the sparse H-polytope.
}
\examples{
library(Matrix)
bineq=c(10,10,10,10,10)
Aineq = matrix(c(1,0,-0.25,-1,2.5,1,0.4,-1,-0.9,0.5), nrow=5, ncol=2, byrow = TRUE)
Aineq = as( Aineq, 'dgCMatrix' )
beq=(0)
Aeq = matrix(, nrow=0, ncol=2, byrow = TRUE)
Aeq=as( Aeq, 'dgCMatrix' )
lb=-100000*c(1,1);
ub=100000*c(1,1);
P <- HpolytopeSparse(Aineq = Aineq, bineq = bineq, Aeq = Aeq, beq = beq, lb = lb, ub = ub)

}
