% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_ball.R
\name{print.simplex_ball_sample}
\alias{print.simplex_ball_sample}
\title{Print a simplex--ball sampling result}
\usage{
\method{print}{simplex_ball_sample}(x, ...)
}
\arguments{
\item{x}{A `simplex_ball_sample` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a simplex--ball sampling result
}
