% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uniform_sampling_correlation_MT}
\alias{uniform_sampling_correlation_MT}
\title{Uniformly sample correlation matrices}
\usage{
uniform_sampling_correlation_MT(n, num_points = 1000, validate = false)
}
\arguments{
\item{n}{The dimension of the correlation matrix.}

\item{num_points}{The number of sample points to generate.}

\item{validate}{Optional. Whether to validate the sampled matrices. Default is false.}
}
\value{
A list of sampled correlation matrices.
}
\description{
Uniformly sample correlation matrices
}