% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VpolytopeClass.R
\docType{class}
\name{Vpolytope-class}
\alias{Vpolytope-class}
\alias{Vpolytope}
\title{An R class to represent a V-polytope}
\description{
A V-polytope is a convex polytope defined by the set of its vertices.
}
\details{
\describe{
   \item{V}{An \eqn{m\times d} numerical matrix that contains the vertices row-wise.}

   \item{volume}{The volume of the polytope if it is known, \eqn{NaN} otherwise by default.}
   
   \item{type}{A character with default value 'Vpolytope', to declare the representation of the polytope.}
}
}
\value{
An object of class \code{Vpolytope} representing the V-polytope.
}
\examples{
V = matrix(c(2,3,-1,7,0,0),ncol = 2, nrow = 3, byrow = TRUE)
P = Vpolytope(V = V)

}
