% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uniform_sample_correlation_matrices}
\alias{uniform_sample_correlation_matrices}
\title{Uniformly sample correlation matrices}
\usage{
uniform_sample_correlation_matrices(
  n,
  num_matrices = 1000L,
  walk_length = 1L,
  nburns = 0L,
  validate = FALSE
)
}
\arguments{
\item{n}{The dimension of the correlation matrix.}

\item{num_matrices}{The number of matrices to generate.}

\item{walk_length}{The length of the random walk.}

\item{nburns}{The number of burn-in steps for the random walk.}

\item{validate}{Optional. Whether to validate the sampled matrices. Default is false.}
}
\value{
A list of sampled correlation matrices.
}
\description{
Uniformly sample correlation matrices
}
